Trait pallet_onboarding::Destroy

pub trait Destroy<AccountId>: Inspect<AccountId> {
    type DestroyWitness;

    fn get_destroy_witness(
        collection: &Self::CollectionId
    ) -> Option<Self::DestroyWitness>;
    fn destroy(
        collection: Self::CollectionId,
        witness: Self::DestroyWitness,
        maybe_check_owner: Option<AccountId>
    ) -> Result<Self::DestroyWitness, DispatchError>;
}

Trait for providing the ability to destroy collections of nonfungible items.

Required Associated Types

type DestroyWitness

The witness data needed to destroy an item.

Required Methods

fn get_destroy_witness(
collection: &Self::CollectionId
) -> Option<Self::DestroyWitness>

Provide the appropriate witness data needed to destroy an item.

fn destroy(
collection: Self::CollectionId,
witness: Self::DestroyWitness,
maybe_check_owner: Option<AccountId>
) -> Result<Self::DestroyWitness, DispatchError>

Destroy an existing fungible item.

collection: The CollectionId to be destroyed.

witness: Any witness data that needs to be provided to complete the operation successfully.

maybe_check_owner: An optional account id that can be used to authorize the destroy command. If not provided, we will not do any authorization checks before destroying the item.

If successful, this function will return the actual witness data from the destroyed item. This may be different than the witness data provided, and can be used to refund weight.