Trait pallet_onboarding::InspectEnumerable

pub trait InspectEnumerable<AccountId>: Inspect<AccountId> {
    fn collections(
    ) -> Box<dyn Iterator<Item = Self::CollectionId> + 'static, Global>;
    fn items(
        collection: &Self::CollectionId
    ) -> Box<dyn Iterator<Item = Self::ItemId> + 'static, Global>;
    fn owned(
        who: &AccountId
    ) -> Box<dyn Iterator<Item = (Self::CollectionId, Self::ItemId)> + 'static, Global>;
    fn owned_in_collection(
        collection: &Self::CollectionId,
        who: &AccountId
    ) -> Box<dyn Iterator<Item = Self::ItemId> + 'static, Global>;
}

Interface for enumerating items in existence or owned by a given account over many collections of NFTs.

Required Methods

fn collections(
) -> Box<dyn Iterator<Item = Self::CollectionId> + 'static, Global>

Returns an iterator of the collections in existence.

fn items(
collection: &Self::CollectionId
) -> Box<dyn Iterator<Item = Self::ItemId> + 'static, Global>

Returns an iterator of the items of a collection in existence.

fn owned(
who: &AccountId
) -> Box<dyn Iterator<Item = (Self::CollectionId, Self::ItemId)> + 'static, Global>

Returns an iterator of the items of all collections owned by who.

fn owned_in_collection(
collection: &Self::CollectionId,
who: &AccountId
) -> Box<dyn Iterator<Item = Self::ItemId> + 'static, Global>

Returns an iterator of the items of collection owned by who.