Trait pallet_onboarding::Mutate

pub trait Mutate<AccountId>: Inspect<AccountId> {
    fn mint_into(
        _collection: &Self::CollectionId,
        _item: &Self::ItemId,
        _who: &AccountId
    ) -> Result<(), DispatchError> { ... }
    fn burn(
        _collection: &Self::CollectionId,
        _item: &Self::ItemId,
        _maybe_check_owner: Option<&AccountId>
    ) -> Result<(), DispatchError> { ... }
    fn set_attribute(
        _collection: &Self::CollectionId,
        _item: &Self::ItemId,
        _key: &[u8],
        _value: &[u8]
    ) -> Result<(), DispatchError> { ... }
    fn set_typed_attribute<K, V>(
        collection: &Self::CollectionId,
        item: &Self::ItemId,
        key: &K,
        value: &V
    ) -> Result<(), DispatchError>
   where
        K: Encode,
        V: Encode,
    { ... }
    fn set_collection_attribute(
        _collection: &Self::CollectionId,
        _key: &[u8],
        _value: &[u8]
    ) -> Result<(), DispatchError> { ... }
    fn set_typed_collection_attribute<K, V>(
        collection: &Self::CollectionId,
        key: &K,
        value: &V
    ) -> Result<(), DispatchError>
   where
        K: Encode,
        V: Encode,
    { ... }
}

Trait for providing an interface for multiple collections of NFT-like items which may be minted, burned and/or have attributes set on them.

Provided Methods

fn mint_into(
_collection: &Self::CollectionId,
_item: &Self::ItemId,
_who: &AccountId
) -> Result<(), DispatchError>

Mint some item of collection to be owned by who.

By default, this is not a supported operation.

fn burn(
_collection: &Self::CollectionId,
_item: &Self::ItemId,
_maybe_check_owner: Option<&AccountId>
) -> Result<(), DispatchError>

Burn some item of collection.

By default, this is not a supported operation.

fn set_attribute(
_collection: &Self::CollectionId,
_item: &Self::ItemId,
_key: &[u8],
_value: &[u8]
) -> Result<(), DispatchError>

Set attribute value of item of collection’s key.

By default, this is not a supported operation.

fn set_typed_attribute<K, V>(
collection: &Self::CollectionId,
item: &Self::ItemId,
key: &K,
value: &V
) -> Result<(), DispatchError>where
K: Encode,
V: Encode,

Attempt to set the strongly-typed attribute value of item of collection’s key.

By default this just attempts to use set_attribute.

fn set_collection_attribute(
_collection: &Self::CollectionId,
_key: &[u8],
_value: &[u8]
) -> Result<(), DispatchError>

Set attribute value of collection’s key.

By default, this is not a supported operation.

fn set_typed_collection_attribute<K, V>(
collection: &Self::CollectionId,
key: &K,
value: &V
) -> Result<(), DispatchError>where
K: Encode,
V: Encode,

Attempt to set the strongly-typed attribute value of collection’s key.

By default this just attempts to use set_attribute.