Trait pallet_onboarding::StaticLookup

pub trait StaticLookup {
    type Source: Codec + Clone + PartialEq<Self::Source> + Debug + TypeInfo;
    type Target;

    fn lookup(s: Self::Source) -> Result<Self::Target, LookupError>;
    fn unlookup(t: Self::Target) -> Self::Source;
}

Means of changing one type into another in a manner dependent on the source type. This variant is different to Lookup in that it doesn’t (can cannot) require any context.

Required Associated Types

type Source: Codec + Clone + PartialEq<Self::Source> + Debug + TypeInfo

Type to lookup from.

type Target

Type to lookup into.

Required Methods

fn lookup(s: Self::Source) -> Result<Self::Target, LookupError>

Attempt a lookup.

fn unlookup(t: Self::Target) -> Self::Source

Convert from Target back to Source.